CF 1.6
- New: IAggregateVectors
- New: Block processing

+++ Catalano.Image/src/Catalano/Imaging/Texture/BinaryPattern/LocalAdaptiveTernaryPattern.java
@@ -81,7 +81,7 @@ public class LocalAdaptiveTernaryPattern implements IBinaryPattern{
      * Initialize a new instance of the LocalTernaryPattern class.
      * @param threshold Threshold.
      */
-    public LocalAdaptiveTernaryPattern(int constant){
+    public LocalAdaptiveTernaryPattern(double constant){
         this.constant = constant;
     }
